requirements/base.txt
Django==5.0.1
djangorestframework==3.14.0
requests==2.31.0
django-filter==23.3
django-cors-headers==4.3.1
django-phonenumber-field==7.2.0
djangorestframework-simplejwt==5.3.0
phonenumbers==8.13.26
Pillow==10.1.0
psycopg==3.1.8
psycopg2-binary==2.9.9
django-dotenv==1.4.2
drf-yasg==1.21.7
